requirements.txt
python-telegram-bot>=20.0
python-dotenv
aiofiles
apscheduler
pytz
httpx
